--- src/main/kotlin/cz.lukynka.prettylog/Main.kt
@@ -1,5 +1,0 @@
-package cz.lukynka.prettylog
-
-fun main() {
-    LoggerFileWriter.load()
-}
